pyproject.toml
[project]
name = "sms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.10",
    "djangorestframework>=3.16.1",
]
